Adjust configurations to fix bootloader command-line
The unlocked bootloader for versions of the V20 other than the H918
misreports the model name as "US996".  With the search and replace
functionality this can be fixed, so do so.  This bootloader can also be
used for the H918 and adds extra functionality, so cover it too.

Investigation has shown certain pieces of functionality are genuinely
changed by the command-line, so having it be correct is important.